Separate out user settings and default to hidden ready for settings menu
