+++ src/index.html
@@ -21,51 +21,51 @@
 		<div class="keys">
 			<div data-key="65" class="white key">
 				<div data-key="87" class="black key">
-					<div class="sound">W</div>
-					<div class="note">C#</div>
+					<div class="sound hidden">W</div>
+					<div class="note hidden">C♯<br>D♭</div>
 				</div>
-				<div class="sound">A</div>
-				<div class="note">C</div>
+				<div class="sound hidden">A</div>
+				<div class="note hidden">C</div>
 			</div>
 			<div data-key="83" class="white key">
 				<div data-key="69" class="black key">
-					<div class="sound">E</div>
-					<div class="note">D#</div>
+					<div class="sound hidden">E</div>
+					<div class="note hidden">D♯<br>E♭</div>
 				</div>
-				<div class="sound">S</div>
-				<div class="note">D</div>
+				<div class="sound hidden">S</div>
+				<div class="note hidden">D</div>
 			</div>
 			<div data-key="68" class="white key">
-				<div class="sound">D</div>
-				<div class="note">E</div>
+				<div class="sound hidden">D</div>
+				<div class="note hidden">E</div>
 			</div>
 			<div data-key="70" class="white key">
 				<div data-key="84" class="black key">
-					<div class="sound">T</div>
-					<div class="note">F#</div>
+					<div class="sound hidden">T</div>
+					<div class="note hidden">F♯<br>G♭</div>
 				</div>
-				<div class="sound">F</div>
-				<div class="note">F</div>
+				<div class="sound hidden">F</div>
+				<div class="note hidden">F</div>
 			</div>
 			<div data-key="71" class="white key">
 				<div data-key="89" class="black key">
-					<div class="sound">Y</div>
-					<div class="note">G#</div>
+					<div class="sound hidden">Y</div>
+					<div class="note hidden">G♯<br>A♭</div>
 				</div>
-				<div class="sound">G</div>
-				<div class="note">G</div>
+				<div class="sound hidden">G</div>
+				<div class="note hidden">G</div>
 			</div>
 			<div data-key="72" class="white key">
 				<div data-key="85" class="black key">
-					<div class="sound">U</div>
-					<div class="note">A#</div>
+					<div class="sound hidden">U</div>
+					<div class="note hidden">A♯<br>B♭</div>
 				</div>
-				<div class="sound">H</div>
-				<div class="note">A</div>
+				<div class="sound hidden">H</div>
+				<div class="note hidden">A</div>
 			</div>
 			<div data-key="74" class="white key">
-				<div class="sound">J</div>
-				<div class="note">B</div>
+				<div class="sound hidden">J</div>
+				<div class="note hidden">B</div>
 			</div>
 		</div>
 	</section>
