+++ src/index.js
@@ -1,6 +1,9 @@
 import { play } from './js/play';
 import randomNoteGenerator from './js/libs/randomNoteGenerator';
+import { userSettings } from './js/userSettings';
+
 import './scss/main.scss'
 
 randomNoteGenerator();
 play();
+userSettings();
