+++ src/js/play.js
@@ -52,33 +52,6 @@ function play() {
 			key.classList.add('fail');
 		}
 	}
-
-	document.getElementById('checkbox2').addEventListener ('click', function (e) {
-		const notes = [...(document.getElementsByClassName('note'))];
-		if(e.target.checked) {
-			notes.forEach(i => {
-				i.classList.add('hidden');
-			});
-		} else {
-			notes.forEach(i => {
-				i.classList.remove('hidden');
-			});
-		}
-	});
-
-	document.getElementById('checkbox1').addEventListener ('click', function (e) {
-		const letters = [...(document.getElementsByClassName('sound'))];
-		if(e.target.checked) {
-			letters.forEach(i => {
-				i.classList.add('hidden');
-			});
-		} else {
-			letters.forEach(i => {
-				i.classList.remove('hidden');
-			});
-		}
-	});
-
 }
 
 export { play }
